4/18 1.map对象类里增加.tmx文件解析函数 2.引入py脚本解码base64，以及解压zlib数据 3.MapSprite不再作为一个单独的组件类，而是集成到map对象类中，由map类进行统一绘制 4.可以对原来的地图放大渲染

+++ RPG_Frame/include/Component/AnimSpriteComponent.h
@@ -15,7 +15,7 @@ namespace GameFrame {
 		float mAnimFPS;//动画播放速率，一秒钟放多少帧
 		int textureW;
 		int textureH;
-		Anim(SDL_Texture* tex,int frame,bool isplay,Vector2 pos,int fps, int w, int h):
+		Anim(SDL_Texture* tex,int frame,bool isplay,Vector2 pos,float fps, int w, int h):
 			mAnimTextures(tex),
 			frame(frame),
 			IsLoopPlay(isplay),
